third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/iterationComposite-expected.txt

This is a testharness.js-based test.
PASS iterationComposite of discrete type animation (align-content)
FAIL iterationComposite of <length> type animationassert_equals: Animated margin-left style at 0s of the third iteration expected "20px" but got "0px"
FAIL iterationComposite of <percentage> type animationassert_equals: Animated width style at 0s of the third iteration expected "100px" but got "0px"
FAIL iterationComposite of <color> type animationassert_equals: Animated color style at 0s of the third iteration expected "rgb(240, 240, 240)" but got "rgb(0, 0, 0)"
FAIL iterationComposite of <color> type animation that green component is decreasingassert_equals: Animated color style at 0s of the third iteration expected "rgb(120, 240, 120)" but got "rgb(0, 120, 0)"
FAIL iterationComposite of <number> type animationassert_equals: Animated flex-grow style at 0s of the third iteration expected "20" but got "0"
FAIL iterationComposite of <shape> type animationassert_equals: Animated clip style at 50s of the first iteration expected "rect(5px, 5px, 5px, 5px)" but got "rect(5px 5px 5px 5px)"
FAIL iterationComposite of <calc()> value animationassert_equals: Animated calc width style at 0s of the third iteration expected "20px" but got "0px"
FAIL iterationComposite of <calc()> value animation that the values can'tbe reducedassert_equals: Animated calc width style at 0s of the third iteration expected "40px" but got "0px"
FAIL iterationComposite of opacity animationassert_equals: Animated opacity style at 0s of the third iteration expected "0.8" but got "0"
FAIL iterationComposite of box-shadow animationassert_equals: Animated box-shadow style at 0s of the third iteration expected "rgb(240, 240, 240) 20px 20px 20px 0px" but got "rgb(0, 0, 0) 0px 0px 0px 0px"
FAIL iterationComposite of filter blur animationassert_equals: Animated filter blur style at 0s of the third iteration expected "blur(20px)" but got "blur(0px)"
FAIL iterationComposite of filter brightness for different unit animationassert_equals: Animated filter brightness style at 0s of the third iteration expected "brightness(2.6)" but got "brightness(1)"
PASS iterationComposite of filter brightness animation
FAIL iterationComposite of filter drop-shadow animationassert_equals: Animated filter drop-shadow style at 0s of the third iteration expected "drop-shadow(rgb(240, 240, 240) 20px 20px 20px)" but got "drop-shadow(rgb(0, 0, 0) 0px 0px 0px)"
FAIL iterationComposite of same filter list animationassert_equals: Animated filter list at 0s of the third iteration expected "brightness(3) contrast(3)" but got "brightness(1) contrast(1)"
FAIL iterationComposite of discrete filter list because of mismatch of the orderassert_equals: Animated filter list at 50s of the third iteration expected "contrast(4) brightness(4)" but got "contrast(2) brightness(2)"
FAIL iterationComposite of different length filter list animationassert_equals: Animated filter list at 0s of the third iteration expected "sepia(2) contrast(3)" but got "sepia(0)"
PASS iterationComposite of transform(rotate) animation
PASS iterationComposite of transform: [ scale(0), scale(1) ] animation
FAIL iterationComposite of transform: [ scale(1), scale(2) ] animationassert_approx_equals: expected matrix(3, 0, 0, 3, 0, 0) but got matrix(1, 0, 0, 1, 0, 0): Animated transform(scale) style at 0s of the third iteration expected 3 +/- 0.0001 but got 1
FAIL iterationComposite of transform: scale(2) animationassert_approx_equals: expected matrix(2, 0, 0, 2, 0, 0) but got matrix(0, 0, 0, 0, 0, 0): Animated transform(scale) style at 0s of the third iteration expected 2 +/- 0.0001 but got 0
FAIL iterationComposite of transform list animationassert_approx_equals: expected matrix(1, 0, 0, 1, 20, 0) but got matrix(1, 0, 0, 1, 0, 0): Animated transform list at 0s of the third iteration expected 20 +/- 0.0001 but got 0
FAIL iterationComposite of transform of matrix functionassert_approx_equals: expected matrix(6, 0, 0, 6, 60, 0) but got matrix(2, 0, 0, 2, 0, 0): Animated transform of matrix function at 0s of the third iteration expected 6 +/- 0.0001 but got 2
FAIL iterationComposite of transform list animation whose order is mismatchedassert_approx_equals: expected matrix(6, 0, 0, 6, 60, 0) but got matrix(2, 0, 0, 2, 0, 0): Animated transform list at 0s of the third iteration expected 6 +/- 0.0001 but got 2
FAIL iterationComposite of transform list animation whose order is mismatched because of missing functionsassert_approx_equals: expected matrix(3, 0, 0, 3, 40, 0) but got matrix(1, 0, 0, 1, 0, 0): Animated transform list at 0s of the third iteration expected 3 +/- 0.0001 but got 1
FAIL iterationComposite of transform from none to translateassert_approx_equals: expected matrix(1, 0, 0, 1, 20, 0) but got matrix(1, 0, 0, 1, 0, 0): Animated transform list at 0s of the third iteration expected 20 +/- 0.0001 but got 0
FAIL iterationComposite of transform of matrix3d functionassert_approx_equals: expected matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 130, 1) but got matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 30, 1): Animated transform of matrix3d function at 0s of the third iteration expected 130 +/- 0.0001 but got 30
FAIL iterationComposite of transform of rotate3d functionassert_equals: dimension of the matrix: Animated transform of rotate3d function at 0s of the third iteration expected 16 but got 6
FAIL iterationComposite starts with non-zero value animationassert_equals: Animated margin-left style at 0s of the third iteration expected "50px" but got "10px"
FAIL iterationComposite with negative final value animationassert_equals: Animated margin-left style at 0s of the third iteration expected "-10px" but got "10px"
FAIL interationComposite changesassert_equals: Animated style at 50s of the third iteration expected "25px" but got "5px"
FAIL duration changes with iterationComposite(accumulate)assert_equals: Animated style at 50s of the third iteration expected "25px" but got "5px"
Harness: the test ran to completion.
